webapp/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy[asyncio]
psycopg2-binary
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
bcrypt>=3.2.0,<4.0.0
asyncpg